requirements.txt
fastapi
uvicorn
pymupdf
python-docx
langchain
langchain-community
faiss-cpu
python-dotenv
google-generativeai
